Implement Long Polling Asynchronous Requests
Each of the main GET endpoints /user /game and /lobby now has its own
/userpoll /lobbypoll and /gamepoll endpoint. These can be used to
asynchronously get results from the game /user /game and /lobby
entities. They return an object if their respective entity changes in
the database.

The workflow to subscribe to asynchronous requests is simple. For
asynchronously getting a user object with the user id 1:

1. Subscribe to the endpoint:

/userpoll/1/subscribe

2. Get the resource asynchronously:

/userpoll/1

An http request to this endpoint is only returned if something in the
data changes. If nothing changes within 60 seconds, the http request
returns with status code 503 unavailable instead.

3. Once done, unsubscribe from the resource:

/userpoll/1/unsubscribe